editoast: move OccurrenceId and TrainId to views
These types handle API representation of train occurrences,
not database persistence, so they belong in the views layer.
